make Page and Shift keys latching toggles
Avoids two-finger gestures! LEDs blink red when active.
* Page modifier resets on any other key press; Shift persists
* ContextMenu appears with Shift key state
* make ContextMenu non modal (let events fall thru)
* internal function to clear modifier key state

after more refactoring, it will be the responsibility of various actions
to turn the modifier state back off, or not... in the case of Shift to
(de)select individual steps, it should be left alone. When switching
pages, probably should clear it.